ad pytest.ini
