+++ frontends/torch-frontend/torch-frontend/python/test/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+markers =
+    mhlo_tools: marks tests depend on mhlo_tools (deselect with '-m "not mhlo_tools"')
+    attention_rewriter: marks tests which run attention rewriting (deselect with '-m "not attention_rewriter"')
+    serial
